restructure files and add docstrings
